+++ .rollup.cache/C/DATADISK/DATA/Projects/AMC/AMCTSLIB/vue-store-oidc/src/main.js
@@ -1,5 +1,0 @@
-import createPiniaStore from "./stores/create-pinia-store";
-import createRouterMiddleware from "./router/create-router-middleware";
-import { dispatchCustomBrowserEvent } from "./services/browser-event";
-export default { createPiniaStore, createRouterMiddleware, dispatchCustomBrowserEvent };
-//# sourceMappingURL=main.js.map