+++ .rollup.cache/C/DATADISK/DATA/Projects/AMC/AMCTSLIB/vue-store-oidc/src/types/pinia-store.js
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=pinia-store.js.map