+++ .rollup.cache/C/DATADISK/DATA/Projects/AMC/AMCTSLIB/vue-store-oidc/test/hello-world.js
@@ -1,2 +1,0 @@
-export const hello = () => 'Hello world!';
-//# sourceMappingURL=hello-world.js.map